[#804] Update sidebar component
